feat(plugin-sdk): load the admin's rich-text nodes from another surface
B-19 needs the page builder's canvas to edit rich text with the SAME node
registry the admin's field editor uses. Lexical recognises content by the
identity of the classes that wrote it, so an editor built on a different set
reads existing content as PLAIN TEXT — silently, at read time, on documents that
already saved.

## Async, and it must stay async

The node classes bring Lexical and PrismJS with them. Measured on this package:
630KB against a 777KB main bundle, which is why `FieldRenderer` already reaches
the editor through `lazy(() => import(...))` and says so.

Static exports of the classes would have made that eager for every consumer of
`@nextlyhq/admin`, including those who never open a rich-text field, and would
have collapsed the SSR boundary the lazy load exists for. So the surface is a
FUNCTION that imports rather than a value that is imported.

Verified rather than assumed: the main bundle grew by 188 bytes and still
contains zero Lexical internals. tsup hoisted Lexical into a SHARED chunk
(461KB) that both the field editor and the loader import — one copy, and the
field editor's own chunk now holds none. The split is enforced by the bundler
rather than by convention.

It is also the shape PB-D11 asks for — ONE shared, lazily-mounted instance — so
the performance constraint and the feature requirement agree.

## One list, not two

`RICH_TEXT_NODES` and the theme move into `rich-text-kit.ts`, and
`RichTextInput` consumes them. Two lists agree the day they are written; the day
one gains a node is the day a document written in one surface renders as an
unknown block in the other, and neither side raises anything — an unregistered
node is a rendering outcome, not an error.

## No duplicate-copy alarm, deliberately

Lexical already throws on this exact path. `createEditor()` rejects a node class
that is not a `LexicalNode` subclass OF THE COPY THAT EDITOR USES, in production
builds as well as development, and names the class, its type and the Lexical
version. A wrapper of ours would restate that or fire where Lexical's does not,
and a second alarm that disagrees with the real one is worse than none.

Entered `@experimental`, named exports only. The plugin-surface snapshot gains
exactly two entries: `loadRichTextEditorKit (value)` and
`RichTextEditorKit (type)`.
